requirements.txt
# Dyslexic AI - LLM Clone for Neurodivergent Users
# Core Dependencies

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# ML/AI
torch==2.1.0
transformers==4.35.0
sentence-transformers==2.2.2
openai==1.3.0

# NLP
nltk==3.8.1
spacy==3.7.2
textstat==0.7.3

# Data Processing
numpy==1.24.3
pandas==2.1.3

# Async Support
asyncio==3.4.3
aiohttp==3.9.0

# Database
sqlalchemy==2.0.23
alembic==1.13.0

# Caching
redis==5.0.1

# Security
cryptography==41.0.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Monitoring
prometheus-client==0.19.0
sentry-sdk==1.38.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Configuration
pyyaml==6.0.1
python-dotenv==1.0.0

# Utilities
python-multipart==0.0.6
email-validator==2.1.0
